Added volume setting to persist last used volume between sessions. Songs can now seek to fractional seconds

+++ JamBit/MusicPlayer.cs
@@ -62,7 +62,7 @@ namespace JamBit
             return player.playState == WMPPlayState.wmppsPlaying;
         }
 
-        public static void SeekTo(int seconds)
+        public static void SeekTo(double seconds)
         {
             player.controls.currentPosition = seconds;
         }
